[pong] refactor to use "Scenes"
I did this so I can add a menu later. I also gave control of collision handling to the `Court` scene and made the GameOver part a Scene. I'm not sure if this was worth doing tbh. Also: add local scoped aliases and fix `AstNode.IfElse` so it calls `SetParent` on the `else_statements` and make `AstNode.Return.GenIl` use the function's return_type rather than the block type